Migrate bankier data to big decimal

+++ app/src/main/java/com/put/ubi/model/UnitValueWithTime.kt
@@ -1,10 +1,11 @@
 package com.put.ubi.model
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class UnitValueWithTime(
     @SerializedName("y")
-    val value: Float,
+    val value: BigDecimal,
     @SerializedName("x")
-    val time: Float,
+    val time: BigDecimal,
 )
